Moved Gradebooks into their own directory for organizational purposes, renamed `GradeBook` to `BaseGradeBook` for clarity, setup project to not include gdbk files (what we're storing our gradebooks in)
