+++ GradeBook/GradeBook.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Folder Include="StudentRecords\" />
+    <None Remove="*.gdbk" />
   </ItemGroup>
 
   <ItemGroup>
