+++ GradeBook/GradeBooks/RankedGradeBook.cs
@@ -3,13 +3,13 @@ using System.Linq;
 
 using GradeBook.Enums;
 
-namespace GradeBook
+namespace GradeBook.GradeBooks
 {
-    public class StandardGradeBook : GradeBook
+    public class RankedGradeBook : BaseGradeBook
     {
-        public StandardGradeBook(string name, bool isWeighted) : base(name, isWeighted)
+        public RankedGradeBook(string name, bool isWeighted) : base(name, isWeighted)
         {
-            Type = GradeBookType.Standard;
+            Type = GradeBookType.Ranked;
         }
 
         public override void CalculateStatistics()
@@ -23,7 +23,7 @@ namespace GradeBook
             var honorPoints = 0d;
             var duelEnrolledPoints = 0d;
 
-            foreach(var student in Students)
+            foreach (var student in Students)
             {
                 student.LetterGrade = GetLetterGrade(student.AverageGrade);
                 student.GPA = GetGPA(student.LetterGrade, IsWeighted, student.Type);
@@ -62,20 +62,20 @@ namespace GradeBook
             }
 
             //#todo refactor into it's own method with calculations performed here
-            Console.WriteLine("Average Grade of all students is " + (allStudentsPoints/Students.Count));
-            if(campusPoints != 0)
+            Console.WriteLine("Average Grade of all students is " + (allStudentsPoints / Students.Count));
+            if (campusPoints != 0)
                 Console.WriteLine("Average for only local students is " + (campusPoints / Students.Where(e => e.Enrollment == EnrollmentType.Campus).Count()));
-            if(statePoints != 0)
+            if (statePoints != 0)
                 Console.WriteLine("Average for only state students (excluding local) is " + (statePoints / Students.Where(e => e.Enrollment == EnrollmentType.State).Count()));
-            if(nationalPoints != 0)
+            if (nationalPoints != 0)
                 Console.WriteLine("Average for only national students (excluding state and local) is " + (nationalPoints / Students.Where(e => e.Enrollment == EnrollmentType.National).Count()));
-            if(internationalPoints != 0)
+            if (internationalPoints != 0)
                 Console.WriteLine("Average for only international students is " + (internationalPoints / Students.Where(e => e.Enrollment == EnrollmentType.International).Count()));
-            if(standardPoints != 0)
+            if (standardPoints != 0)
                 Console.WriteLine("Average for students excluding honors and duel enrollment is " + (standardPoints / Students.Where(e => e.Type == StudentType.Standard).Count()));
-            if(honorPoints != 0)
+            if (honorPoints != 0)
                 Console.WriteLine("Average for only honors students is " + (honorPoints / Students.Where(e => e.Type == StudentType.Honors).Count()));
-            if(duelEnrolledPoints != 0)
+            if (duelEnrolledPoints != 0)
                 Console.WriteLine("Average for only duel enrolled students is " + (duelEnrolledPoints / Students.Where(e => e.Type == StudentType.DuelEnrolled).Count()));
         }
 
@@ -88,7 +88,7 @@ namespace GradeBook
             Console.WriteLine(student.Name + " (" + student.LetterGrade + ":" + student.AverageGrade + "): GPA: " + student.GPA + ".");
             Console.WriteLine();
             Console.WriteLine("Grades:");
-            foreach(var grade in student.Grades)
+            foreach (var grade in student.Grades)
             {
                 Console.WriteLine(grade);
             }
@@ -98,26 +98,26 @@ namespace GradeBook
         {
             if (isWeighted)
             {
-                switch(studentType)
+                switch (studentType)
                 {
                     case StudentType.DuelEnrolled:
                     case StudentType.Honors:
-                    {
-                        switch(letterGrade)
                         {
-                            case 'A':
-                                return 5;
-                            case 'B':
-                                return 4;
-                            case 'C':
-                                return 3;
-                            case 'D':
-                                return 2;
-                            case 'F':
-                                return 1;
+                            switch (letterGrade)
+                            {
+                                case 'A':
+                                    return 5;
+                                case 'B':
+                                    return 4;
+                                case 'C':
+                                    return 3;
+                                case 'D':
+                                    return 2;
+                                case 'F':
+                                    return 1;
+                            }
+                            break;
                         }
-                        break;
-                    }
                 }
             }
 
@@ -139,13 +139,16 @@ namespace GradeBook
 
         public override char GetLetterGrade(double averageGrade)
         {
-            if (averageGrade >= 90)
+            var grades = Students.OrderByDescending(e => e.AverageGrade).Select(e => e.AverageGrade).ToList();
+            var gradeScale = Students.Count()*0.2;
+
+            if (grades[(int)Math.Ceiling(gradeScale)-1] <= averageGrade)
                 return 'A';
-            else if (averageGrade >= 80)
+            else if (grades[(int)Math.Ceiling(gradeScale*2)-1] <= averageGrade)
                 return 'B';
-            else if (averageGrade >= 70)
+            else if (grades[(int)Math.Ceiling(gradeScale*3)-1] <= averageGrade)
                 return 'C';
-            else if (averageGrade >= 60)
+            else if (grades[(int)Math.Ceiling(gradeScale*4)-1] <= averageGrade)
                 return 'D';
             else
                 return 'F';
