+++ GradeBookTests/TestGradeBook.cs
@@ -5,7 +5,7 @@ namespace GradeBookTests
 {
     //This class exists to provide a default implimentation of the GradeBook for testing
     
-    public class TestGradeBook : GradeBook.GradeBook
+    public class TestGradeBook : GradeBook.BaseGradeBook
     {
         public TestGradeBook(string name, bool isWeighted) : base(name, isWeighted) { }
